LayoutTests/imported/w3c/web-platform-tests/html/semantics/popups/popup-light-dismiss.tentative-expected.txt

CONSOLE MESSAGE: TypeError: popup7.showPopUp is not a function. (In 'popup7.showPopUp()', 'popup7.showPopUp' is undefined)
CONSOLE MESSAGE: TypeError: popup7.showPopUp is not a function. (In 'popup7.showPopUp()', 'popup7.showPopUp' is undefined)
Popup 1 Popup 1 Popup1 anchor (no action) Outside all popups
Inside popup 1  Popup 2 Inside popup 1 after button
Inside popup 2
Next control after popup1  Popup 3 - button 3
Inside popup 4
Inside popup 3
Popup 3 - button 4
Inside popup 5
Inside popup 6
Bottom of popup6
Popup 6
Button Inside popup 8 after button
Popup8 anchor (no action)
Button Inside popup 9 after button
Popup 9
Popup 1
Anchor  Open Popup 2
Popup 4

Popup 2  Open Popup 3
Popup 3  Open Popup 4
Open convoluted popup
Popup
Hint
Manual

Harness Error (FAIL), message = TypeError: popup7.showPopUp is not a function. (In 'popup7.showPopUp()', 'popup7.showPopUp' is undefined)

FAIL Clicking outside a popup will dismiss the popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Clicking inside a popup does not close that popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Moving focus outside the popup will dismiss the popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Clicking inside a child popup shouldn't close either popup promise_test: Unhandled rejection with value: object "TypeError: popup1.showPopUp is not a function. (In 'popup1.showPopUp()', 'popup1.showPopUp' is undefined)"
FAIL Clicking inside a parent popup should close child popup promise_test: Unhandled rejection with value: object "TypeError: popup1.showPopUp is not a function. (In 'popup1.showPopUp()', 'popup1.showPopUp' is undefined)"
FAIL Clicking on invoking element, after using it for activation, shouldn't close its popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Clicking on invoking element, even if it wasn't used for activation, shouldn't close its popup promise_test: Unhandled rejection with value: object "TypeError: popup1.showPopUp is not a function. (In 'popup1.showPopUp()', 'popup1.showPopUp' is undefined)"
FAIL Clicking on popuptoggletarget element, even if it wasn't used for activation, should hide it exactly once promise_test: Unhandled rejection with value: object "TypeError: popup1.showPopUp is not a function. (In 'popup1.showPopUp()', 'popup1.showPopUp' is undefined)"
FAIL Clicking on anchor element (that isn't an invoking element) shouldn't close its popup promise_test: Unhandled rejection with value: object "TypeError: popup1.showPopUp is not a function. (In 'popup1.showPopUp()', 'popup1.showPopUp' is undefined)"
FAIL Dragging from an open popup outside an open popup should leave the popup open promise_test: Unhandled rejection with value: object "TypeError: popup1.showPopUp is not a function. (In 'popup1.showPopUp()', 'popup1.showPopUp' is undefined)"
FAIL An invoking element should be part of the ancestor chain promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL An invoking element that was not used to invoke the popup can still be part of the ancestor chain promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Scrolling within a popup should not close the popup promise_test: Unhandled rejection with value: object "TypeError: popup6.showPopUp is not a function. (In 'popup6.showPopUp()', 'popup6.showPopUp' is undefined)"
FAIL Clicking inside a shadow DOM popup does not close that popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Clicking outside a shadow DOM popup should close that popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Moving focus back to the anchor element should not dismiss the popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Moving focus back to the active trigger element should not dismiss the popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Moving focus back to an inactive trigger element should also *not* dismiss the popup promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Ensure circular/convoluted ancestral relationships are functional promise_test: Unhandled rejection with value: object "TypeError: convPopup1.showPopUp is not a function. (In 'convPopup1.showPopUp()', 'convPopup1.showPopUp' is undefined)"
FAIL Ensure circular/convoluted ancestral relationships are functional, with a direct showPopUp() promise_test: Unhandled rejection with value: object "TypeError: convPopup1.showPopUp is not a function. (In 'convPopup1.showPopUp()', 'convPopup1.showPopUp' is undefined)"
FAIL Light dismiss of mixed popup types promise_test: Unhandled rejection with value: object "TypeError: auto.showPopUp is not a function. (In 'auto.showPopUp()', 'auto.showPopUp' is undefined)"
